Refactor NIC schema and expanders in VMM v2: Consolidate NIC handling by introducing `nic_backing_info` and `network_info` fields, while deprecating legacy fields. Enhance schema definitions and expanders for better clarity and maintainability. This update aims to streamline the NIC configuration process and improve the overall structure of the VMM v2 service.
