+++ nutanix/services/vmmv2/resource_nutanix_ova_vm_deploy_v2.go
@@ -54,171 +54,7 @@ func ResourceNutanixOvaVMDeploymentV2() *schema.Resource {
 						"nics": {
 							Type:     schema.TypeList,
 							Required: true,
-							Elem: &schema.Resource{
-								Schema: map[string]*schema.Schema{
-									"ext_id": {
-										Type:     schema.TypeString,
-										Computed: true,
-									},
-									"backing_info": {
-										Type:     schema.TypeList,
-										Optional: true,
-										Elem: &schema.Resource{
-											Schema: map[string]*schema.Schema{
-												"model": {
-													Type:     schema.TypeString,
-													Optional: true,
-												},
-												"mac_address": {
-													Type:     schema.TypeString,
-													Optional: true,
-												},
-												"is_connected": {
-													Type:     schema.TypeBool,
-													Optional: true,
-												},
-												"num_queues": {
-													Type:     schema.TypeInt,
-													Optional: true,
-													Default:  1,
-												},
-											},
-										},
-									},
-									"network_info": {
-										Type:     schema.TypeList,
-										Optional: true,
-										Elem: &schema.Resource{
-											Schema: map[string]*schema.Schema{
-												"nic_type": {
-													Type:     schema.TypeString,
-													Optional: true,
-													ValidateFunc: validation.StringInSlice([]string{
-														"SPAN_DESTINATION_NIC",
-														"NORMAL_NIC", "DIRECT_NIC", "NETWORK_FUNCTION_NIC",
-													}, false),
-												},
-												"network_function_chain": {
-													Type:     schema.TypeList,
-													Optional: true,
-													Elem: &schema.Resource{
-														Schema: map[string]*schema.Schema{
-															"ext_id": {
-																Type:     schema.TypeString,
-																Optional: true,
-															},
-														},
-													},
-												},
-												"network_function_nic_type": {
-													Type:     schema.TypeString,
-													Optional: true,
-													ValidateFunc: validation.StringInSlice([]string{
-														"TAP", "EGRESS",
-														"INGRESS",
-													}, false),
-												},
-												"subnet": {
-													Type:     schema.TypeList,
-													Optional: true,
-													Elem: &schema.Resource{
-														Schema: map[string]*schema.Schema{
-															"ext_id": {
-																Type:     schema.TypeString,
-																Optional: true,
-															},
-														},
-													},
-												},
-												"vlan_mode": {
-													Type:         schema.TypeString,
-													Optional:     true,
-													ValidateFunc: validation.StringInSlice([]string{"TRUNK", "ACCESS"}, false),
-												},
-												"trunked_vlans": {
-													Type:     schema.TypeList,
-													Optional: true,
-													Elem: &schema.Schema{
-														Type: schema.TypeInt,
-													},
-												},
-												"should_allow_unknown_macs": {
-													Type:     schema.TypeBool,
-													Optional: true,
-												},
-												"ipv4_config": {
-													Type:     schema.TypeList,
-													Optional: true,
-													Elem: &schema.Resource{
-														Schema: map[string]*schema.Schema{
-															"should_assign_ip": {
-																Type:     schema.TypeBool,
-																Optional: true,
-															},
-															"ip_address": {
-																Type:     schema.TypeList,
-																Optional: true,
-																Elem: &schema.Resource{
-																	Schema: map[string]*schema.Schema{
-																		"value": {
-																			Type:     schema.TypeString,
-																			Optional: true,
-																		},
-																		"prefix_length": {
-																			Type:     schema.TypeInt,
-																			Optional: true,
-																		},
-																	},
-																},
-															},
-															"secondary_ip_address_list": {
-																Type:     schema.TypeList,
-																Optional: true,
-																Elem: &schema.Resource{
-																	Schema: map[string]*schema.Schema{
-																		"value": {
-																			Type:     schema.TypeString,
-																			Optional: true,
-																		},
-																		"prefix_length": {
-																			Type:     schema.TypeInt,
-																			Optional: true,
-																		},
-																	},
-																},
-															},
-														},
-													},
-												},
-												"ipv4_info": {
-													Type:     schema.TypeList,
-													Optional: true,
-													Elem: &schema.Resource{
-														Schema: map[string]*schema.Schema{
-															"learned_ip_addresses": {
-																Type:     schema.TypeList,
-																Optional: true,
-																Elem: &schema.Resource{
-																	Schema: map[string]*schema.Schema{
-																		"value": {
-																			Type:     schema.TypeString,
-																			Required: true,
-																		},
-																		"prefix_length": {
-																			Type:     schema.TypeInt,
-																			Optional: true,
-																		},
-																	},
-																},
-															},
-														},
-													},
-												},
-											},
-										},
-									},
-								},
-							},
+							Elem:     nicsElemSchemaV2(),
 						},
 						"cd_roms": {
 							Type:     schema.TypeList,
