+++ nutanix/services/vmmv2/resource_nutanix_template_deploy_v2.go
@@ -8,7 +8,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 	import2 "github.com/nutanix/ntnx-api-golang-clients/prism-go-client/v4/models/prism/v4/config"
 	import4 "github.com/nutanix/ntnx-api-golang-clients/vmm-go-client/v4/models/common/v1/config"
 	import1 "github.com/nutanix/ntnx-api-golang-clients/vmm-go-client/v4/models/prism/v4/config"
@@ -68,177 +67,7 @@ func ResourceNutanixTemplateDeployV2() *schema.Resource {
 						"nics": {
 							Type:     schema.TypeList,
 							Optional: true,
-							Elem: &schema.Resource{
-								Schema: map[string]*schema.Schema{
-									"ext_id": {
-										Type:     schema.TypeString,
-										Computed: true,
-									},
-									"backing_info": {
-										Type:     schema.TypeList,
-										Optional: true,
-										Elem: &schema.Resource{
-											Schema: map[string]*schema.Schema{
-												"model": {
-													Type:     schema.TypeString,
-													Optional: true,
-												},
-												"mac_address": {
-													Type:     schema.TypeString,
-													Optional: true,
-												},
-												"is_connected": {
-													Type:     schema.TypeBool,
-													Optional: true,
-												},
-												"num_queues": {
-													Type:     schema.TypeInt,
-													Optional: true,
-													Default:  1,
-												},
-											},
-										},
-									},
-									"network_info": {
-										Type:     schema.TypeList,
-										Optional: true,
-										Elem: &schema.Resource{
-											Schema: map[string]*schema.Schema{
-												"nic_type": {
-													Type:     schema.TypeString,
-													Optional: true,
-													ValidateFunc: validation.StringInSlice([]string{
-														"SPAN_DESTINATION_NIC",
-														"NORMAL_NIC", "DIRECT_NIC", "NETWORK_FUNCTION_NIC",
-													}, false),
-												},
-												"network_function_chain": {
-													Type:     schema.TypeList,
-													Optional: true,
-													Elem: &schema.Resource{
-														Schema: map[string]*schema.Schema{
-															"ext_id": {
-																Type:     schema.TypeString,
-																Optional: true,
-															},
-														},
-													},
-												},
-												"network_function_nic_type": {
-													Type:     schema.TypeString,
-													Optional: true,
-													ValidateFunc: validation.StringInSlice([]string{
-														"TAP", "EGRESS",
-														"INGRESS",
-													}, false),
-												},
-												"subnet": {
-													Type:     schema.TypeList,
-													Optional: true,
-													Elem: &schema.Resource{
-														Schema: map[string]*schema.Schema{
-															"ext_id": {
-																Type:     schema.TypeString,
-																Optional: true,
-															},
-														},
-													},
-												},
-												"vlan_mode": {
-													Type:         schema.TypeString,
-													Optional:     true,
-													ValidateFunc: validation.StringInSlice([]string{"TRUNK", "ACCESS"}, false),
-												},
-												"trunked_vlans": {
-													Type:     schema.TypeList,
-													Optional: true,
-													Elem: &schema.Schema{
-														Type: schema.TypeInt,
-													},
-												},
-												"should_allow_unknown_macs": {
-													Type:     schema.TypeBool,
-													Optional: true,
-												},
-												"ipv4_config": {
-													Type:     schema.TypeList,
-													Optional: true,
-													Elem: &schema.Resource{
-														Schema: map[string]*schema.Schema{
-															"should_assign_ip": {
-																Type:     schema.TypeBool,
-																Optional: true,
-															},
-															"ip_address": {
-																Type:     schema.TypeList,
-																Optional: true,
-																Elem: &schema.Resource{
-																	Schema: map[string]*schema.Schema{
-																		"value": {
-																			Type:     schema.TypeString,
-																			Optional: true,
-																		},
-																		"prefix_length": {
-																			Type:     schema.TypeInt,
-																			Optional: true,
-																			Default:  defaultValue,
-																		},
-																	},
-																},
-															},
-															"secondary_ip_address_list": {
-																Type:     schema.TypeList,
-																Optional: true,
-																Elem: &schema.Resource{
-																	Schema: map[string]*schema.Schema{
-																		"value": {
-																			Type:     schema.TypeString,
-																			Optional: true,
-																		},
-																		"prefix_length": {
-																			Type:     schema.TypeInt,
-																			Optional: true,
-																			Default:  defaultValue,
-																		},
-																	},
-																},
-															},
-														},
-													},
-												},
-												// not visible in API reference
-												"ipv4_info": {
-													Type:     schema.TypeList,
-													Optional: true,
-													Computed: true,
-													Elem: &schema.Resource{
-														Schema: map[string]*schema.Schema{
-															"learned_ip_addresses": {
-																Type:     schema.TypeList,
-																Optional: true,
-																Computed: true,
-																Elem: &schema.Resource{
-																	Schema: map[string]*schema.Schema{
-																		"value": {
-																			Type:     schema.TypeString,
-																			Required: true,
-																		},
-																		"prefix_length": {
-																			Type:     schema.TypeInt,
-																			Optional: true,
-																			Default:  defaultValue,
-																		},
-																	},
-																},
-															},
-														},
-													},
-												},
-											},
-										},
-									},
-								},
-							},
+							Elem:     nicsElemSchemaV2(),
 						},
 						"guest_customization": schemaForGuestCustomization(),
 					},
@@ -458,198 +287,6 @@ func expandNic(pr []interface{}) []config.Nic {
 		return nicList
 	}
 	return nil
-}
-
-// expandVirtualEthernetNic expands either:
-// - nic_backing_info.virtual_ethernet_nic (VirtualEthernetNic), OR
-// - nic_network_info.virtual_ethernet_nic_network_info (VirtualEthernetNicNetworkInfo).
-//
-// It returns the concrete (non-pointer) SDK object to satisfy OneOf setters, or nil if empty.
-func expandVirtualEthernetNic(pr interface{}) interface{} {
-	if pr == nil {
-		return nil
-	}
-	list, ok := pr.([]interface{})
-	if !ok || len(list) == 0 || list[0] == nil {
-		return nil
-	}
-	val, ok := list[0].(map[string]interface{})
-	if !ok {
-		return nil
-	}
-
-	// Backing object uses model/mac/is_connected/num_queues keys.
-	if _, hasModel := val["model"]; hasModel || val["mac_address"] != nil || val["num_queues"] != nil || val["is_connected"] != nil {
-		ven := config.NewVirtualEthernetNic()
-
-		if model, ok := val["model"]; ok && model != nil && model.(string) != "" {
-			ven.Model = common.ExpandEnum[config.VirtualEthernetNicModel](model.(string))
-		}
-		if mac, ok := val["mac_address"]; ok && mac != nil && mac.(string) != "" {
-			ven.MacAddress = utils.StringPtr(mac.(string))
-		}
-		if isConn, ok := val["is_connected"]; ok && isConn != nil {
-			ven.IsConnected = utils.BoolPtr(isConn.(bool))
-		}
-		if nq, ok := val["num_queues"]; ok && nq != nil {
-			ven.NumQueues = utils.IntPtr(nq.(int))
-		}
-		return *ven
-	}
-
-	// Otherwise treat as network info object.
-	venNI := config.NewVirtualEthernetNicNetworkInfo()
-
-	// nic_type (optional; defaults to NORMAL_NIC server-side, but we infer to avoid "missing NIC" diffs)
-	if nicType, ok := val["nic_type"]; ok && nicType != nil && nicType.(string) != "" {
-		venNI.NicType = common.ExpandEnum[config.NicType](nicType.(string))
-	}
-	if ntwkFunc, ok := val["network_function_chain"]; ok && ntwkFunc != nil {
-		venNI.NetworkFunctionChain = expandNetworkFunctionChainReference(ntwkFunc)
-	}
-	if ntwkFuncNicType, ok := val["network_function_nic_type"]; ok && ntwkFuncNicType != nil && ntwkFuncNicType.(string) != "" {
-		venNI.NetworkFunctionNicType = common.ExpandEnum[config.NetworkFunctionNicType](ntwkFuncNicType.(string))
-	}
-	if subnet, ok := val["subnet"]; ok && subnet != nil {
-		venNI.Subnet = expandSubnetReference(subnet)
-	}
-	if vlanMode, ok := val["vlan_mode"]; ok && vlanMode != nil && vlanMode.(string) != "" {
-		venNI.VlanMode = common.ExpandEnum[config.VlanMode](vlanMode.(string))
-	}
-	if trunkedVlans, ok := val["trunked_vlans"]; ok && trunkedVlans != nil {
-		vlansList := trunkedVlans.([]interface{})
-		trunkInts := make([]int, len(vlansList))
-		for i, v := range vlansList {
-			trunkInts[i] = v.(int)
-		}
-		venNI.TrunkedVlans = trunkInts
-	}
-	if unknownMacs, ok := val["should_allow_unknown_macs"]; ok && unknownMacs != nil {
-		venNI.ShouldAllowUnknownMacs = utils.BoolPtr(unknownMacs.(bool))
-	}
-	if ipv4, ok := val["ipv4_config"]; ok && ipv4 != nil {
-		venNI.Ipv4Config = expandIpv4Config(ipv4)
-	}
-	if ipv4Info, ok := val["ipv4_info"]; ok && ipv4Info != nil {
-		venNI.Ipv4Info = expandIPv4Info(ipv4Info)
-	}
-
-	// Infer nic_type if absent (helps avoid perpetual "nics { ... }" diffs when config sets subnet/ipv4 only).
-	if venNI.NicType == nil {
-		if venNI.NetworkFunctionNicType != nil || venNI.NetworkFunctionChain != nil {
-			p := config.NICTYPE_NETWORK_FUNCTION_NIC
-			venNI.NicType = p.Ref()
-		} else if venNI.Subnet != nil || venNI.Ipv4Config != nil || venNI.VlanMode != nil || len(venNI.TrunkedVlans) > 0 || venNI.ShouldAllowUnknownMacs != nil {
-			p := config.NICTYPE_NORMAL_NIC
-			venNI.NicType = p.Ref()
-		}
-	}
-
-	return *venNI
-}
-
-// expandSriovNic expands either:
-// - nic_backing_info.sriov_nic (SriovNic), OR
-// - nic_network_info.sriov_nic_network_info (SriovNicNetworkInfo).
-func expandSriovNic(pr interface{}) interface{} {
-	if pr == nil {
-		return nil
-	}
-	list, ok := pr.([]interface{})
-	if !ok || len(list) == 0 || list[0] == nil {
-		return nil
-	}
-	val, ok := list[0].(map[string]interface{})
-	if !ok {
-		return nil
-	}
-
-	// Network info is only vlan_id.
-	if _, hasVlanID := val["vlan_id"]; hasVlanID {
-		ni := config.NewSriovNicNetworkInfo()
-		if vlanID, ok := val["vlan_id"]; ok && vlanID != nil {
-			ni.VlanId = utils.IntPtr(vlanID.(int))
-		}
-		return *ni
-	}
-
-	// Backing info.
-	sriov := config.NewSriovNic()
-	if mac, ok := val["mac_address"]; ok && mac != nil && mac.(string) != "" {
-		sriov.MacAddress = utils.StringPtr(mac.(string))
-	}
-	if isConn, ok := val["is_connected"]; ok && isConn != nil {
-		sriov.IsConnected = utils.BoolPtr(isConn.(bool))
-	}
-	if hostRef, ok := val["host_pcie_device_reference"]; ok && hostRef != nil {
-		sriov.HostPcieDeviceReference = expandHostPcieDeviceReference(hostRef)
-	}
-	if prof, ok := val["sriov_profile_reference"]; ok && prof != nil {
-		sriov.SriovProfileReference = expandNicProfileReference(prof)
-	}
-	return *sriov
-}
-
-// expandDpOffloadNic expands either:
-// - nic_backing_info.dp_offload_nic (DpOffloadNic), OR
-// - nic_network_info.dp_offload_nic_network_info (DpOffloadNicNetworkInfo).
-func expandDpOffloadNic(pr interface{}) interface{} {
-	if pr == nil {
-		return nil
-	}
-	list, ok := pr.([]interface{})
-	if !ok || len(list) == 0 || list[0] == nil {
-		return nil
-	}
-	val, ok := list[0].(map[string]interface{})
-	if !ok {
-		return nil
-	}
-
-	// Backing info contains dp_offload_profile_reference / host_pcie_device_reference / mac/is_connected.
-	if _, hasProfile := val["dp_offload_profile_reference"]; hasProfile || val["host_pcie_device_reference"] != nil || val["mac_address"] != nil || val["is_connected"] != nil {
-		dp := config.NewDpOffloadNic()
-		if mac, ok := val["mac_address"]; ok && mac != nil && mac.(string) != "" {
-			dp.MacAddress = utils.StringPtr(mac.(string))
-		}
-		if isConn, ok := val["is_connected"]; ok && isConn != nil {
-			dp.IsConnected = utils.BoolPtr(isConn.(bool))
-		}
-		if hostRef, ok := val["host_pcie_device_reference"]; ok && hostRef != nil {
-			dp.HostPcieDeviceReference = expandHostPcieDeviceReference(hostRef)
-		}
-		if prof, ok := val["dp_offload_profile_reference"]; ok && prof != nil {
-			dp.DpOffloadProfileReference = expandNicProfileReference(prof)
-		}
-		return *dp
-	}
-
-	// Otherwise network info.
-	ni := config.NewDpOffloadNicNetworkInfo()
-	if subnet, ok := val["subnet"]; ok && subnet != nil {
-		ni.Subnet = expandSubnetReference(subnet)
-	}
-	if vlanMode, ok := val["vlan_mode"]; ok && vlanMode != nil && vlanMode.(string) != "" {
-		ni.VlanMode = common.ExpandEnum[config.VlanMode](vlanMode.(string))
-	}
-	if trunkedVlans, ok := val["trunked_vlans"]; ok && trunkedVlans != nil {
-		vlansList := trunkedVlans.([]interface{})
-		trunkInts := make([]int, len(vlansList))
-		for i, v := range vlansList {
-			trunkInts[i] = v.(int)
-		}
-		ni.TrunkedVlans = trunkInts
-	}
-	if unknownMacs, ok := val["should_allow_unknown_macs"]; ok && unknownMacs != nil {
-		ni.ShouldAllowUnknownMacs = utils.BoolPtr(unknownMacs.(bool))
-	}
-	if ipv4, ok := val["ipv4_config"]; ok && ipv4 != nil {
-		ni.Ipv4Config = expandIpv4Config(ipv4)
-	}
-	if ipv4Info, ok := val["ipv4_info"]; ok && ipv4Info != nil {
-		ni.Ipv4Info = expandIPv4Info(ipv4Info)
-	}
-	return *ni
 }
 
 func expandHostPcieDeviceReference(pr interface{}) *config.HostPcieDeviceReference {
