+++ nutanix/services/vmmv2/resource_nutanix_template_v2.go
@@ -1840,193 +1840,7 @@ func schemaForNics() *schema.Schema {
 		Type:     schema.TypeList,
 		Optional: true,
 		Computed: true,
-		Elem: &schema.Resource{
-			Schema: map[string]*schema.Schema{
-				"tenant_id": {
-					Type:     schema.TypeString,
-					Computed: true,
-				},
-				"links": schemaForLinks(),
-				"ext_id": {
-					Type:     schema.TypeString,
-					Computed: true,
-				},
-				"backing_info": {
-					Type:     schema.TypeList,
-					Optional: true,
-					Computed: true,
-					Elem: &schema.Resource{
-						Schema: map[string]*schema.Schema{
-							"model": {
-								Type:         schema.TypeString,
-								Optional:     true,
-								Computed:     true,
-								ValidateFunc: validation.StringInSlice([]string{"VIRTIO", "E1000"}, false),
-							},
-							"mac_address": {
-								Type:     schema.TypeString,
-								Optional: true,
-								Computed: true,
-							},
-							"is_connected": {
-								Type:     schema.TypeBool,
-								Optional: true,
-								Default:  true,
-							},
-							"num_queues": {
-								Type:     schema.TypeInt,
-								Optional: true,
-								Default:  1,
-							},
-						},
-					},
-				},
-				"network_info": {
-					Type:     schema.TypeList,
-					Optional: true,
-					Computed: true,
-					Elem: &schema.Resource{
-						Schema: map[string]*schema.Schema{
-							"nic_type": {
-								Type:         schema.TypeString,
-								Optional:     true,
-								Computed:     true,
-								ValidateFunc: validation.StringInSlice([]string{"SPAN_DESTINATION_NIC", "NORMAL_NIC", "DIRECT_NIC", "NETWORK_FUNCTION_NIC"}, false),
-							},
-							"network_function_chain": {
-								Type:     schema.TypeList,
-								Optional: true,
-								Computed: true,
-								Elem: &schema.Resource{
-									Schema: map[string]*schema.Schema{
-										"ext_id": {
-											Type:     schema.TypeString,
-											Optional: true,
-											Computed: true,
-										},
-									},
-								},
-							},
-							"network_function_nic_type": {
-								Type:         schema.TypeString,
-								Optional:     true,
-								Computed:     true,
-								ValidateFunc: validation.StringInSlice([]string{"TAP", "EGRESS", "INGRESS"}, false),
-							},
-							"subnet": {
-								Type:     schema.TypeList,
-								Optional: true,
-								Computed: true,
-								Elem: &schema.Resource{
-									Schema: map[string]*schema.Schema{
-										"ext_id": {
-											Type:     schema.TypeString,
-											Optional: true,
-											Computed: true,
-										},
-									},
-								},
-							},
-							"vlan_mode": {
-								Type:         schema.TypeString,
-								Optional:     true,
-								Computed:     true,
-								ValidateFunc: validation.StringInSlice([]string{"TRUNK", "ACCESS"}, false),
-							},
-							"trunked_vlans": {
-								Type:     schema.TypeList,
-								Optional: true,
-								Computed: true,
-								Elem: &schema.Schema{
-									Type: schema.TypeInt,
-								},
-							},
-							"should_allow_unknown_macs": {
-								Type:     schema.TypeBool,
-								Optional: true,
-								Computed: true,
-							},
-							"ipv4_config": {
-								Type:     schema.TypeList,
-								Optional: true,
-								Computed: true,
-								Elem: &schema.Resource{
-									Schema: map[string]*schema.Schema{
-										"should_assign_ip": {
-											Type:     schema.TypeBool,
-											Optional: true,
-											Computed: true,
-										},
-										"ip_address": {
-											Type:     schema.TypeList,
-											Computed: true,
-											Elem: &schema.Resource{
-												Schema: map[string]*schema.Schema{
-													"value": {
-														Type:     schema.TypeString,
-														Required: true,
-													},
-													"prefix_length": {
-														Type:     schema.TypeInt,
-														Optional: true,
-														Default:  defaultValue,
-													},
-												},
-											},
-										},
-										"secondary_ip_address_list": {
-											Type:     schema.TypeList,
-											Computed: true,
-											Elem: &schema.Resource{
-												Schema: map[string]*schema.Schema{
-													"value": {
-														Type:     schema.TypeString,
-														Required: true,
-													},
-													"prefix_length": {
-														Type:     schema.TypeInt,
-														Optional: true,
-														Default:  defaultValue,
-													},
-												},
-											},
-										},
-									},
-								},
-							},
-							// not visible in API reference
-							"ipv4_info": {
-								Type:     schema.TypeList,
-								Optional: true,
-								Computed: true,
-								Elem: &schema.Resource{
-									Schema: map[string]*schema.Schema{
-										"learned_ip_addresses": {
-											Type:     schema.TypeList,
-											Optional: true,
-											Computed: true,
-											Elem: &schema.Resource{
-												Schema: map[string]*schema.Schema{
-													"value": {
-														Type:     schema.TypeString,
-														Required: true,
-													},
-													"prefix_length": {
-														Type:     schema.TypeInt,
-														Optional: true,
-														Default:  defaultValue,
-													},
-												},
-											},
-										},
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-		},
+		Elem: nicsElemSchemaV2WithTenantLinks(),
 	}
 }
