+++ nutanix/services/vmmv2/resource_nutanix_vms_clone_v2.go
@@ -61,173 +61,7 @@ func ResourceNutanixVMCloneV2() *schema.Resource {
 				Type:     schema.TypeList,
 				Optional: true,
 				Computed: true,
-				Elem: &schema.Resource{
-					Schema: map[string]*schema.Schema{
-						"ext_id": {
-							Type:     schema.TypeString,
-							Optional: true,
-							Computed: true,
-						},
-						"backing_info": {
-							Type:     schema.TypeList,
-							Optional: true,
-							Computed: true,
-							Elem: &schema.Resource{
-								Schema: map[string]*schema.Schema{
-									"model": {
-										Type:         schema.TypeString,
-										Optional:     true,
-										Computed:     true,
-										ValidateFunc: validation.StringInSlice([]string{"VIRTIO", "E1000"}, false),
-									},
-									"mac_address": {
-										Type:     schema.TypeString,
-										Optional: true,
-										Computed: true,
-									},
-									"is_connected": {
-										Type:     schema.TypeBool,
-										Optional: true,
-										Computed: true,
-									},
-									"num_queues": {
-										Type:     schema.TypeInt,
-										Optional: true,
-										Default:  1,
-									},
-								},
-							},
-						},
-						"network_info": {
-							Type:     schema.TypeList,
-							Optional: true,
-							Computed: true,
-							Elem: &schema.Resource{
-								Schema: map[string]*schema.Schema{
-									"nic_type": {
-										Type:     schema.TypeString,
-										Optional: true,
-										Computed: true,
-										ValidateFunc: validation.StringInSlice([]string{
-											"SPAN_DESTINATION_NIC",
-											"NORMAL_NIC", "DIRECT_NIC", "NETWORK_FUNCTION_NIC",
-										}, false),
-									},
-									"network_function_chain": {
-										Type:     schema.TypeList,
-										Optional: true,
-										Computed: true,
-										Elem: &schema.Resource{
-											Schema: map[string]*schema.Schema{
-												"ext_id": {
-													Type:     schema.TypeString,
-													Optional: true,
-													Computed: true,
-												},
-											},
-										},
-									},
-									"network_function_nic_type": {
-										Type:     schema.TypeString,
-										Optional: true,
-										Computed: true,
-										ValidateFunc: validation.StringInSlice([]string{
-											"TAP", "EGRESS",
-											"INGRESS",
-										}, false),
-									},
-									"subnet": {
-										Type:     schema.TypeList,
-										Optional: true,
-										Computed: true,
-										Elem: &schema.Resource{
-											Schema: map[string]*schema.Schema{
-												"ext_id": {
-													Type:     schema.TypeString,
-													Optional: true,
-													Computed: true,
-												},
-											},
-										},
-									},
-									"vlan_mode": {
-										Type:         schema.TypeString,
-										Optional:     true,
-										Computed:     true,
-										ValidateFunc: validation.StringInSlice([]string{"TRUNK", "ACCESS"}, false),
-									},
-									"trunked_vlans": {
-										Type:     schema.TypeList,
-										Optional: true,
-										Computed: true,
-										Elem: &schema.Schema{
-											Type: schema.TypeInt,
-										},
-									},
-									"should_allow_unknown_macs": {
-										Type:     schema.TypeBool,
-										Optional: true,
-										Computed: true,
-									},
-									"ipv4_config": {
-										Type:     schema.TypeList,
-										Optional: true,
-										Computed: true,
-										Elem: &schema.Resource{
-											Schema: map[string]*schema.Schema{
-												"should_assign_ip": {
-													Type:     schema.TypeBool,
-													Optional: true,
-													Computed: true,
-												},
-												"ip_address": {
-													Type:     schema.TypeList,
-													Optional: true,
-													Computed: true,
-													Elem: &schema.Resource{
-														Schema: map[string]*schema.Schema{
-															"value": {
-																Type:     schema.TypeString,
-																Optional: true,
-																Computed: true,
-															},
-															"prefix_length": {
-																Type:     schema.TypeInt,
-																Optional: true,
-																Computed: true,
-															},
-														},
-													},
-												},
-												"secondary_ip_address_list": {
-													Type:     schema.TypeList,
-													Optional: true,
-													Computed: true,
-													Elem: &schema.Resource{
-														Schema: map[string]*schema.Schema{
-															"value": {
-																Type:     schema.TypeString,
-																Optional: true,
-																Computed: true,
-															},
-															"prefix_length": {
-																Type:     schema.TypeInt,
-																Optional: true,
-																Computed: true,
-															},
-														},
-													},
-												},
-											},
-										},
-									},
-									// not visible in API reference
-									// "ipv4Info":   {},
-								},
-							},
-						},
-					},
-				},
+				Elem:     nicsElemSchemaV2(),
 			},
 			"boot_config": {
 				Type:     schema.TypeList,
